+++ otserv/trunk/items.h
@@ -38,9 +38,9 @@ public:
 	unsigned short id;
 
   unsigned short maxItems;   // maximum size if this is a container
-	unsigned short weight;     // weight of the item, e.g. throwing distance depends on it
-	std::string    name;		   // the name of the item
-	std::string description;	// additional description... as in "The blade is a magic flame." for fireswords
+	double weight;						 // weight of the item, e.g. throwing distance depends on it
+	std::string    name;			 // the name of the item
+	std::string description;	 // additional description... as in "The blade is a magic flame." for fireswords
   WeaponType     weaponType;
   int            attack;
   int            defence;
